oskelly-material-mapping-GL/requirements.txt
openai>=1.57.0
pandas>=2.0.0
PyYAML>=6.0
openpyxl>=3.1.5
requests>=2.31.0

